git-svn-id: http://rdinnovation.onf.fr/svn/plugin-lvox@11 f1561d03-89af-ab4a-a97c-2bf9dc1f41eb

+++ pluginlvox/tools/lvox_computegridtools.h
@@ -91,10 +91,10 @@ private:
                               QVector3D &inPoint,
                               QVector3D &outPoint);
 
-    static bool updateIntervals(const float &bc,
-                         const float &uc,
-                         const float &sc,
-                         const float &sd,
+    static bool updateIntervals(const double &bc,
+                         const double &uc,
+                         const double &sc,
+                         const double &sd,
                          double &t0, double &t1);
